Rewrite internal test tool + add basic integer tests
This changes the test command to use absolute paths instead of relative, and then clean them up when we're printing them out to the user.
We're also now using the fully Zen language evaluator so we can actually run and test or tests in a real environment.
Additionally a few basic tests has been added to get the test suite starter.