Create basic bytecode compiler
Right now we only support integers, however, we're able to turn integer literals found in the AST into integer objects, and then push them onto a stack in our compiler as constants so we can use the later.
There are also some helper methods both for running the compiler tests, and for deconstructing the bytecodes, so we're able to stringify the instructions so it's easier to debug in the future.