Fix bytecode supporting compiled functions serialization
This also changes how the code CLI command works by calling String() directly on the bytecode struct, this allows us to access the constants while formatting the buffered output, so we're able to include the inner instructions within compiled functions in the output, as well as the scope index.

The tests for the bytecode changes have also been updated to reflect these changes.