Implement builtins for the VM
This refactors a big part of how builtins are created and used within every part of the language engine, which allows us to reuse the functions for compilation and within the VM at runtime. Every builtin now works everywhere, however, the only thing that's missing is being able to capture the output of print builtins so we're able to test against the data, and so the existing VM tests doesn't print test data while running.