Add support for using compiled functions as args for builtins
This fixes the issue of some builtins such as array.filter not working in the compiler/VM since they take a function as an argument that is called to evaluate to some value within the VM itself, the VM and evaluator now produces the same result.

There may be ways to go about executing the compiled functions without creating a new VM instance, while also cleaning up the stack and picking up the return value like a normal executed function, but also supporting recursively calling the functions, however, for now this works, and work for optimizing the VM can be done at a later time.