Add support for combing output types in debug command
When a script being tested both prints something to STDOUT and also causes an error the error will be omitted unless it's being run in verbose state, however even without that it would be nice to see the error so we can compare the evaluator and VM as intended. This now work more in line with expectations.