Add serialized testing to the VM test runner
This adds a step that first serializes, and then deserializes the the bytecode, and then runs the deserialized bytecode with the VM test runner as well to ensure we still meet all our expectations.

Additionally some new timings have been added to track serialization and deserialization when running in verbose mode.