Fix parser correctly stopping suffix expressions
If a suffix expression such as `i++` or `i--` were used without a semicolon at the end of a inner body, such as a function or loop, it would incorrectly take up the next token, which closed the body, causing things outside of the function or loop to be included within the inner body.
This should now be fixed since we're checking if the next token is actually a semicolon, and if it's not we don't continue to the next token, this change also makes it possible to do something like `println(i++)` in the evaluator, however, the compiler/vm will need to be updated to support the new syntax.