Add check for suffix kind at compile time
This adds symbol kinds for numbers, strings, and booleans, which will allow us to check more types/kinds at compile time, it also adds a check to ensure we're working with numbers when we're trying to increment or decrement values.
Because the symbol kind is set for both variable definitions and reassignments, we can be pretty sure we're working with a number value.