Fix reassigning mutable variable in conditional outside of root scope
Previously if you tried to reassign a mutable variable inside a conditional inside a function/closure without any other instructions, and the conditional is followed by a OpCall instruction it would incorrectly get the value of the OpCall, it would print the "builtin function" string, instead of the value itself, this only happened with calls and no returns, nor if the condition contained more than a single assignment.
This issue should be fixed now that we're pushing the value of the symbol back on the stack after setting it to a new value.